Complete

+++ SuperMarioBros3/BoomerangBrother.cpp
@@ -133,7 +133,7 @@ void CBoomerangBrother::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 	else
 	{
 		float min_tx, min_ty;
-		int nx = 0, ny = 0;
+		float nx = 0, ny = 0;
 		float rdx = 0;
 		float rdy = 0;
