+++ SuperMarioBros3/IntroObject.cpp
@@ -55,7 +55,7 @@ void CIntroObject::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 	{
 
 		float min_tx, min_ty, x0, y0;
-		int nx = 0, ny = 0;
+		float nx = 0, ny = 0;
 		float rdx = 0;
 		float rdy = 0;
