+++ SuperMarioBros3/WorldPlayer.cpp
@@ -55,7 +55,7 @@ void CWorldPlayer::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 		else
 		{
 			float min_tx, min_ty;
-			int nx = 0, ny = 0;
+			float nx = 0, ny = 0;
 			float rdx = 0;
 			float rdy = 0;
